Fix: better completion saving
